Migrate ConnectionsScene state management to ViewModel
- Move @State values from ConnectionsScene to ConnectionsViewModel
- Migrate isPresentingScanner and isPresentingAlertMessage to ViewModel
- Move all action methods (onScan, onPaste, onHandleScan, onSelectDisconnect) to ViewModel
- Move connectURI private method to ViewModel for centralized error handling
- Update Scene to use @State private var model pattern with bindings
- Remove redundant Actions extension from Scene
- Add UIKit import to ViewModel for UIPasteboard access
- Follow ImportWalletScene pattern for state management
