+++ Features/WalletConnector/Sources/WalletConnector/ViewModels/ConnectionsViewModel.swift
@@ -1,6 +1,7 @@
 // Copyright (c). Gem Wallet. All rights reserved.
 
 import Foundation
+import UIKit
 import Primitives
 import Store
 import Localization
@@ -15,12 +16,16 @@ public final class ConnectionsViewModel {
     
     var request: ConnectionsRequest
     var connections: [WalletConnection] = []
+    var isPresentingScanner: Bool = false
+    var isPresentingAlertMessage: AlertMessage?
 
     public init(
         service: ConnectionsService
     ) {
         self.service = service
         self.request = ConnectionsRequest()
+        self.isPresentingScanner = false
+        self.isPresentingAlertMessage = nil
     }
 
     var title: String { Localized.WalletConnect.title }
@@ -61,5 +66,49 @@ public final class ConnectionsViewModel {
     
     func updateSessions() {
         service.updateSessions()
+    }
+}
+
+// MARK: - Actions
+
+extension ConnectionsViewModel {
+    func onScan() {
+        isPresentingScanner = true
+    }
+    
+    func onPaste() {
+        guard let content = UIPasteboard.general.string else {
+            return
+        }
+
+        Task {
+            await connectURI(uri: content)
+        }
+    }
+    
+    func onHandleScan(_ result: String) {
+        Task {
+            await connectURI(uri: result)
+        }
+    }
+    
+    func onSelectDisconnect(_ connection: WalletConnection) {
+        Task {
+            do {
+                try await disconnect(connection: connection)
+            } catch {
+                isPresentingAlertMessage = AlertMessage(message: error.localizedDescription)
+                NSLog("disconnect error: \(error)")
+            }
+        }
+    }
+    
+    private func connectURI(uri: String) async {
+        do {
+            try await pair(uri: uri)
+        } catch {
+            isPresentingAlertMessage = AlertMessage(message: error.localizedDescription)
+            NSLog("connectURI error: \(error)")
+        }
     }
 }
